WIP: Added code for video animation
Added code to register the scroll position and the video element to
calculate and apply the delay on the video when scrolling.

BUG: delay is not defined (even thoug it seems to be in the code) also
I get the error "HTMLMediaElement.currentTime setter: Value being assigned is not a finite floating-point value" when trying to set the playback time of the video to the delay time.